Added input handling, game object removal, score incrementation.
Basically, the game is playable now. Input system is too complicated for it's own sake though, could've been much easier if I used groups instead of separate class...

To do next:
death animations, increasing difficulties, backgrounds, sounds, GUI, better images, penalties for spamming mouseclicks (accuracy).

+++ src/com/vincas/miceandcheese/components/Attack.java
@@ -3,17 +3,17 @@ package com.vincas.miceandcheese.components;
 import com.artemis.Component;
 
 public class Attack extends Component {
-	private int attackDamage;
+	private float attackDamage;
 
-	public Attack(int attackDamage) {
+	public Attack(float attackDamage) {
 		this.attackDamage = attackDamage;
 	}
 
-	public void setAttackDamage(int attackDamage) {
+	public void setAttackDamage(float attackDamage) {
 		this.attackDamage = attackDamage;
 	}
 
-	public int getAttackDamage() {
+	public float getAttackDamage() {
 		return attackDamage;
 	}
 }
